[Product] [Dashboard] integrate with dropzone

+++ database/migrations/2020_11_06_143353_create_categories_table.php
@@ -17,7 +17,7 @@ class CreateCategoriesTable extends Migration
             $table->id();
             $table->string('name');
             $table->string('slug');
-            $table->integer('parent_id')->nullable();
+            $table->smallInteger('parent_id')->nullable();
             $table->timestamps();
             $table->softDeletes();
         });
